requirements.txt
torch
torchvision
nibabel
timm
Pillow
numpy
matplotlib
torchmetrics
lpips